+++ Source/ImoutoRebirth.Lilin/ImoutoRebirth.Lilin.Application/TagSlice/SetTagAliasesCommand.cs
@@ -1,0 +1,5 @@
+﻿using ImoutoRebirth.Common.Cqrs.Abstract;
+
+namespace ImoutoRebirth.Lilin.Application.TagSlice;
+
+public record SetTagAliasesCommand(Guid TagId, IReadOnlyCollection<Guid> AliasTagIds) : ICommand;